% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{add_reg_folder}
\alias{add_reg_folder}
\title{Tell nat.templatebrains about local folder containing registrations}
\usage{
add_reg_folder(dir, first = TRUE)
}
\arguments{
\item{dir}{Path to one or more folders containing registrations (Please see
\bold{File layout} section for details)}

\item{first}{Whether the new folder should be added to the start (default) or
end of the search list.}
}
\description{
This sets options('nat.templatebrains.regdirs') appropriately so
  that registrations can be found by e.g. \code{xform_brain}.
}
\section{File layout}{
 You must pass a folder containing one or more
  registrations, not the registration folder itself. So if you have this
  situation on disk \itemize{

  \item myregistrations/

  \item myregistrations/reg1.list

  \item myregistrations/reg2.list

  }

  you will be passing \code{add_reg_folder("/path/to/registrations")}
}
\examples{
\dontrun{
  add_reg_folder("myextraregistrations")
}
# addiing a non-existent folder will generate an error
tools::assertError(add_reg_folder(tempfile()))
}

