% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.character.templatebrain}
\alias{as.character.templatebrain}
\title{Convert template brain to character vector representation}
\usage{
\method{as.character}{templatebrain}(x, field = c("regName", "name"))
}
\arguments{
\item{x}{templatebrain to convert}

\item{field}{Which field to use (defaults to \code{'regName'})}
}
\value{
character vector
}
\description{
This will normally be used to extract the short name i.e. \code{regName}.
}

