% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_slice.R
\name{display_slice}
\alias{display_slice}
\title{Display an image slice in 3D}
\usage{
display_slice(brain, slice, ...)
}
\arguments{
\item{brain}{template brain (e.g. IS2) of the slice.}

\item{slice}{Path to PNG image containing slice to display.}

\item{...}{extra arguments to pass to \code{\link[rgl]{persp3d}}.}
}
\description{
Display an image slice in 3D
}
