% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE.R
\docType{data}
\name{FCWB.demo}
\alias{FCWB.demo}
\title{Sample template brain: FlyCircuit Whole Brain}
\description{
This is a sample template brain for testing purposes which is equivalent to
the FCWB template brain defined by the \code{nat.flybrains}, which should be
considered the canonical version.
}
