% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-xform.R
\name{fit_xform}
\alias{fit_xform}
\title{Fit affine or thin plate spline transform to arbitrary transformation}
\usage{
fit_xform(
  reg,
  samplepts,
  refpts = NULL,
  type = c("affine", "rigid", "similarity", "tps"),
  subsample = FALSE,
  ...
)
}
\arguments{
\item{reg}{Any registration compatible with \code{nat::\link{xform}},
including non-rigid and multi-step registrations. You must either supply
this or the \code{refpts} argument.}

\item{samplepts}{A set of points in the sample (floating) space. Can be any
object compatible with \code{nat::\link{xyzmatrix}}.}

\item{refpts}{An optional set of points in the target (fixed) space matching
\code{samplepts}. You must either supply this or the \code{reg} argument.}

\item{type}{A character string specifying the type of registration. See
\code{\link{computeTransform}} for details.}

\item{subsample}{A number of points to subsample from
\code{samplepts,refpts}. The default value of \code{FALSE} means use all
provided points to calculate the new transform.}

\item{...}{Additional arguments passed to \code{\link{computeTransform}}}
}
\value{
A homogeneous affine matrix or a \code{\link{tpsreg}} object
}
\description{
Fit affine or thin plate spline transform to arbitrary transformation
}
\examples{
\dontrun{
library(nat.flybrains)
reg=shortest_bridging_seq(sample='FCWB', reference="JFRC2")
fit_xform(reg, nat::kcs20, subsample=200, type='affine')
}
}
