% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{nat.templatebrains-package}
\alias{nat.templatebrains}
\alias{nat.templatebrains-package}
\title{NeuroAnatomy Toolbox add-on package for handling template brains}
\description{
This package provides a class \code{templatebrain} that stores key
information about referennce brains along with helpfer functions to simplify
transformation of data between template brains (a.k.a bridging) and mirroring
of data within a template brain. Presently there
}
\section{Helper functions}{
 Easy-to-use functions for transforming data from
  one template brain to another, displaying slices alongside 3D data, etc.
  are provided. See especially \code{\link{xform_brain}},
  \code{\link{mirror_brain}} and \code{\link{plot3d.templatebrain}}.
}

\section{Package options}{


  \itemize{

  \item options('nat.templatebrains.extrabridge') specifies a character vector of
  additional directories containing bridging registrations.

  \item options('nat.templatebrains.extramirror') specifies a character vector of
  additional directories containing mirroring registrations.

  }

  Note that registrations in these extra directories will override those of
  the same name provided with the package.
}
\seealso{
\code{\link[nat]{nat}}
}
\keyword{package}
\keyword{registration}
\keyword{template}

