% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3d.templatebrain}
\alias{plot3d.templatebrain}
\title{Plot 3D surface of a template brain}
\usage{
\method{plot3d}{templatebrain}(x, col = "grey", alpha = 0.3, ...)
}
\arguments{
\item{x}{the template brain to plot.}

\item{col}{the color of the surface.}

\item{alpha}{the alpha value of the surface.}

\item{...}{extra arguments to pass to \code{\link[rgl]{plot3d}}.}
}
\description{
Plot 3D surface of a template brain
}
\details{
This function will work immediately for the standard
  \link{template-brains} defined in the package documentation. If passed an
  object called e.g. FCWB it expects to find another object named FCWB.surf
  containing the surface information. If you follow this naming convention
  for user-defined refbrains it will work for them as well.
}
\seealso{
\code{\link{template-brains}}
}

