% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.templatebrain}
\alias{print.templatebrain}
\title{Print brain template information in human-readable form}
\usage{
\method{print}{templatebrain}(x, ...)
}
\arguments{
\item{x}{the object of class \code{templatebrain} to print.}

\item{...}{further objects to print.}
}
\description{
Print brain template information in human-readable form
}

