% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.templatebrain}
\alias{as.templatebrain}
\alias{as.templatebrain.character}
\alias{as.templatebrain.im3d}
\title{Use image file or other object to initialise template brain}
\usage{
as.templatebrain(x, ...)

\method{as.templatebrain}{character}(x, name = NULL, ...)

\method{as.templatebrain}{im3d}(x, name, ...)
}
\arguments{
\item{x}{object used to construct the templatebrain, either a character
vector with the path to a file or a \code{im3d} object.}

\item{...}{additional named arguments passed to methods or to
\code{templatebrain} and that will be added as fields to the
\code{templatebrain} object.}

\item{name}{name of the template brain. Will use the filename (minus final
extension) by default for \code{as.templatebrain.character} but must be
supplied for \code{as.templatebrain.im3d}.}
}
\value{
A list with class \code{templatebrain}.
}
\description{
Use image file or other object to initialise template brain
}
\seealso{
\code{\link[nat]{im3d}}
}

