% Generated by roxygen2 (4.0.2): do not edit by hand
\name{display_slice}
\alias{display_slice}
\title{Display an image slice in 3D}
\usage{
display_slice(brain, slice, ...)
}
\arguments{
\item{brain}{template brain (e.g. IS2) of the slice.}

\item{slice}{path to the slice PNG image to display.}

\item{...}{extra arguments to pass to persp3d.}
}
\description{
Display an image slice in 3D
}

