% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{local_reg_dir_for_url}
\alias{local_reg_dir_for_url}
\title{Standard local checkout location for extra registration directories}
\usage{
local_reg_dir_for_url(url = NULL)
}
\arguments{
\item{url}{Character vector containing a url. When \code{url=NULL} defaults
to giving the base path.}
}
\description{
When called without any argument returns the root directory that will be
inspected for extra registrations. You can put them there yourself manually,
but you are better off in general using \code{\link{add_reg_repo}} to install
those that can be found in our github repos e.g.
\href{https://github.com/jefferislab/BridgingRegistrations}{jefferislab/BridgingRegistrations}
}
\details{
Note that this location will always be the same place on a machine i.e. this
defines a consistent, persistent location on disk to store data across
sesssions.
}

