% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mirror_brain}
\alias{mirror_brain}
\title{Mirror 3D object around a given axis, optionally using a warping registration}
\usage{
mirror_brain(x, brain, mirrorAxis = c("X", "Y", "Z"), ...)
}
\arguments{
\item{x}{the 3D object to be mirrored.}

\item{brain}{source template brain (e.g. IS2) that data is in.}

\item{mirrorAxis}{the axis to mirror (default \code{"X"}).}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
Mirror 3D object around a given axis, optionally using a warping registration
}

