% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{update_reg_repos}
\alias{update_reg_repos}
\title{Update local copy of git repository containing registrations}
\usage{
update_reg_repos(x = NULL)
}
\arguments{
\item{x}{Path to local checkout of a registration git repository. See details
for meaning of default.}
}
\description{
When \code{x=NULL} all repositories listed in
options(nat.templatebrains.regdirs) are checked to see if they are git
repositories and, if yes, they are pulled to update.
}
\seealso{
\code{\link{add_reg_repo}}
}

