% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xform_brain}
\alias{xform_brain}
\title{Transform 3D object between template brains}
\usage{
xform_brain(x, sample, reference, ...)
}
\arguments{
\item{x}{the 3D object to be transformed}

\item{sample}{source template brain (e.g. IS2) that data is currently in.}

\item{reference}{target template brain (e.g. IS2) that data should be
transformed into.}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
Transform 3D object between template brains
}
\details{
NB the sample and reference brains can either be
  \code{templatebrain} objects or a character string containing the short
  name of the template e.g. \code{"IS2"}.
}

